Moved OnCompletion feature code out of OnCompletion.test.ts to src/task/OnCompletion.ts
